unicode-table-generator: key case-mapping multis by full code point
Widen the `multis` key from u16 to u32 and store the full code point
for U+1DF95, whose title and upper mappings leave plane 1, so the entry
is no longer confused with a plane-relative low.

This does not build yet. The generated lookup still searches with the
u16 `input_low`:

    l2_lut.multis.binary_search_by_key(&input_low, |&(p, _)| p)

which no longer type-checks against the u32 key, and the outputs are
still rebuilt with `reconstruct(input_high, output_low)`, so U+1DF95
would decode to U+10053 rather than U+0053 even once it compiles.

The draft UCD files in src/tools/unicode-table-generator/unicode-downloads/
are left untracked.

+++ src/tools/unicode-table-generator/src/case_mapping.rs
@@ -349,7 +349,7 @@ impl Range {
     }
 }
 
-fn deconstruct(c: char) -> (u16, u16) {
+fn deconstruct(c: char) -> (u32, u16) {
     let c = c as u32;
     let plane = (c >> 16) as u16;
     let low = c as u16;
@@ -391,7 +391,7 @@ fn lookup(input: char, l1_lut: &L1Lut) -> Option<[char; 3]> {
         }
     };
 
-    if let Ok(idx) = l2_lut.multis.binary_search_by_key(&input_low, |&(p, _)| p) {
+    if let Ok(idx) = l2_lut.multis.binary_search_by_key(&input_low as u32, |&(p, _)| p) {
         // SAFETY: binary search guarantees that the index is in bounds.
         let &(_, output_lows) = unsafe { l2_lut.multis.get_unchecked(idx) };
         // SAFETY: Table data are guaranteed to be valid Unicode.
